Move delete site confirm dialog to SiteManagement #6261
